refactor(bluesky): isolate write-safety gates in validation module
The bridge app mixed request models, routes, runner wiring, and the
write-safety validation gates in one module; the gates now live in a
small validation module documenting their invariants (fresh
writes_enabled read at startup, per-launch session re-hash), with
runner wiring and request models split out alongside. Route paths,
response shapes, and gate wiring are unchanged.
